wasm64: preserve user callback pointer width

+++ arch/wasm/include/asm/wasm_imports.h
@@ -54,9 +54,10 @@ int wasm_import(user, compile_end)(u32 maximum_memory_pages);
 void wasm_import(user, compile_abort)(void);
 void wasm_import(user, instantiate)(bool fresh_memory);
 void wasm_import(user, call)(void);
-void wasm_import(user, switch_entry)(u32 fn, u32 arg);
-void wasm_import(user, call_signal_handler)(u32 fn, u32 sig);
-int wasm_import(user, call_siginfo_handler)(u32 trampoline, u32 fn, u32 sig);
+void wasm_import(user, switch_entry)(uintptr_t fn, uintptr_t arg);
+void wasm_import(user, call_signal_handler)(uintptr_t fn, u32 sig);
+int wasm_import(user, call_siginfo_handler)(uintptr_t trampoline,
+					    uintptr_t fn, u32 sig);
 void wasm_import(user, halt_signal_handler)(void);
 
 int wasm_import(user, read)(void *to, const void __user *from, unsigned long n);
